@@ -33,24 +33,27 @@
   "readme": "",
   "readmeFilename": "README.md",
   "dependencies": {
-    "lodash": "^4.17.4"
+    "lodash": "^4.17.5",
+    "reflect-metadata": "^0.1.12",
+    "tslib": "^1.9.0"
   },
   "devDependencies": {
-    "@types/chai": "^3.4.35",
-    "@types/benchmark": "^1.0.30",
-    "@types/lodash": "^4.14.59",
-    "@types/mocha": "^2.2.40",
-    "@types/node": "^7.0.12",
-    "@types/sinon": "^2.1.0",
-    "@types/sinon-chai": "^2.7.27",
+    "@types/benchmark": "^1.0.31",
+    "@types/chai": "^4.1.2",
+    "@types/lodash": "^4.14.106",
+    "@types/mocha": "^5.0.0",
+    "@types/node": "^9.6.2",
+    "@types/sinon": "^4.3.1",
+    "@types/sinon-chai": "^2.7.29",
     "benchmark": "^2.1.4",
-    "chai": "3.5.0",
-    "grunt": "1.0.1",
+    "chai": "4.1.2",
+    "grunt": "1.0.2",
     "grunt-cli": "^1.2.0",
-    "grunt-mocha-test": "0.13.2",
-    "mocha": "^3.2.0",
-    "sinon": "^2.1.0",
-    "sinon-chai": "^2.9.0",
-    "typescript": "^2.4.1"
+    "grunt-mocha-test": "0.13.3",
+    "mocha": "^5.0.5",
+    "sinon": "^4.5.0",
+    "sinon-chai": "^3.0.0",
+    "sleep-promise": "^2.0.0",
+    "typescript": "^2.8.1"
   }
 }
