@@ -1,6 +1,6 @@
 "use strict";
 import chai = require('chai');
-import    inject = require('../lib/inject');
+import    ioc = require('../lib/inject');
 import {Injector} from "../lib/inject";
 
 let should = chai.should();
@@ -10,7 +10,7 @@ describe('Property Array', function () {
         let injector: Injector;
 
         beforeEach(function () {
-            injector = inject.createContainer();
+            injector = ioc.createContainer();
 
             let Rectangle = class {
                 objects: any[]
@@ -136,12 +136,12 @@ describe('Property Array', function () {
             }
 
 
-            injector = inject.createContainer()
+            injector = ioc.createContainer()
 
-            injector.define('rectangle', Rectangle)
+            injector.register('rectangle', Rectangle)
                 .injectArray('objects', [{ref: 'fooManager'}, {ref: 'barManager'}])
-                .define('fooManager', FooManager).singleton()
-                .define('barManager', BarManager).singleton()
+            injector.register('fooManager', FooManager).singleton()
+            injector.register('barManager', BarManager).singleton()
 
 
             injector.initialize()
