@@ -1,8 +1,10 @@
 "use strict";
 import {Injector} from "../lib/inject";
-let should = require('chai').should(),
-    inject = require('../lib/inject');
+import ioc = require('../lib/inject');
+import chai = require('chai');
+import {define, singleton, inject} from "../lib/decorators";
 
+let should = chai.should();
 describe('Singleton', function () {
 
 
@@ -10,7 +12,7 @@ describe('Singleton', function () {
         let injector: Injector;
 
         beforeEach(function () {
-            injector = inject.createContainer();
+            injector = ioc.createContainer();
 
             class Rectangle {
                 number: number
@@ -57,11 +59,59 @@ describe('Singleton', function () {
         });
     });
 
+    describe('create singleton object with decorators', function () {
+        let injector: Injector;
+
+        beforeEach(function () {
+            injector = ioc.createContainer();
+
+            @define()
+            @singleton()
+            class Rectangle {
+                number: number
+
+                constructor() {
+                    this.number = Math.random();
+                }
+
+                area() {
+                    return 25;
+                }
+            }
+
+            injector.register(Rectangle)
+
+            injector.initialize();
+        });
+
+        it('should save object in instances', function () {
+            should.exist(injector.getInstances()['rectangle']);
+        });
+
+        it('should get object', function () {
+
+            let rectangle: any = injector.getObject('rectangle');
+
+            should.exist(rectangle);
+            rectangle.area().should.equal(25);
+        });
+
+        it('should have the same instance ', function () {
+
+            let rectangle: any = injector.getObject('rectangle');
+            let number = rectangle.number;
+            let rectangle2: any = injector.getObject('rectangle');
+
+            number.should.equal(rectangle2.number);
+
+        });
+    });
+
     describe('create not singleton object', function () {
         let injector: Injector;
 
         beforeEach(function () {
-            injector = inject.createContainer();
+            injector = ioc.createContainer();
 
             class Rectangle {
                 number: number
