@@ -27,10 +27,10 @@ suite.add("get object singleton", () => {
     let req1 = injector.getObject('rectangle');
 });
 suite.add("get object singleton ignore factory", () => {
-    let req1 = injector.getObject('rectangle', [], true);
+    let req1 = injector.getObject('rectangle', []);
 });
 suite.add("get object not singleton", () => {
-    let req1 = injector.getObject('rectangle2', [], true);
+    let req1 = injector.getObject('rectangle2', []);
 });
 suite.on('cycle', (event) => {
     console.log(String(event.target));
