@@ -2,31 +2,37 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 const _ = require("lodash");
 class Define {
-    constructor(injector, id, type) {
+    constructor(id, type) {
         if (!id) {
             return;
         }
-        this.id = id;
-        this.injector = injector;
-        this.definition = {
+        this._id = id;
+        this._definition = {
+            id: this._id,
             type: type,
             inject: [],
             alias: [],
             aliasFactory: [],
             args: []
         };
-        injector.addDefinition(id, this.definition);
+    }
+    get definition() {
+        return this._definition;
     }
     type(type) {
-        this.definition.type = type;
+        this._definition.type = type;
         return this;
     }
     singleton(singleton = true) {
-        this.definition.singleton = _.isUndefined(singleton) ? true : singleton;
+        this._definition.singleton = _.isUndefined(singleton) ? true : singleton;
+        return this;
+    }
+    factory(factory = true) {
+        this._definition.factory = _.isUndefined(factory) ? true : factory;
         return this;
     }
     lazy(lazy = true) {
-        this.definition.lazy = _.isUndefined(lazy) ? true : lazy;
+        this._definition.lazy = _.isUndefined(lazy) ? true : lazy;
         return this;
     }
     inject(name, inject) {
@@ -34,16 +40,16 @@ class Define {
             name = name.split(" ");
         }
         if (_.isArray(name)) {
-            this.definition.inject.push.apply(this.definition.inject, name);
+            this._definition.inject.push.apply(this._definition.inject, name);
         }
         else if (_.isObject(name)) {
-            this.definition.inject.push(name);
+            this._definition.inject.push(name);
         }
         else if (_.toArray(arguments).length == 1 && _.isString(name)) {
-            this.definition.inject.push({ name: name, ref: name });
+            this._definition.inject.push({ name: name, ref: name });
         }
         else if (_.toArray(arguments).length == 2 && _.isString(name)) {
-            this.definition.inject.push({ name: name, ref: inject || name });
+            this._definition.inject.push({ name: name, ref: inject || name });
         }
         return this;
     }
@@ -82,7 +88,7 @@ class Define {
     injectFactory(name, factory) {
         return this.inject({
             name: name,
-            factory: factory
+            factory: factory || name + "Factory"
         });
     }
     injectObjectProperty(name, object, propertyName) {
@@ -102,41 +108,38 @@ class Define {
     }
     alias(alias) {
         if (_.isArray(alias)) {
-            this.definition.alias.push.apply(this.definition.alias, alias);
+            this._definition.alias.push.apply(this._definition.alias, alias);
         }
         else {
-            this.definition.alias.push(alias);
+            this._definition.alias.push(alias);
         }
         return this;
     }
     initMethod(initMethod) {
-        this.definition.initMethod = initMethod || "initialize";
+        this._definition.initMethod = initMethod || "initialize";
         return this;
     }
     injectorAware() {
-        this.definition.injectorAware = true;
+        this._definition.injectorAware = true;
         return this;
     }
     aliasFactory(aliasFactory) {
         if (_.isArray(aliasFactory)) {
-            this.definition.aliasFactory.push.apply(this.definition.aliasFactory, aliasFactory);
+            this._definition.aliasFactory.push.apply(this._definition.aliasFactory, aliasFactory);
         }
         else {
-            this.definition.aliasFactory.push(aliasFactory);
+            this._definition.aliasFactory.push(aliasFactory);
         }
         return this;
     }
     args(args) {
         if (_.isArray(args)) {
-            this.definition.args.push.apply(this.definition.args, args);
+            this._definition.args.push.apply(this._definition.args, args);
         }
         else {
-            this.definition.args.push(args);
+            this._definition.args.push(args);
         }
         return this;
-    }
-    define(id, type) {
-        return this.injector.define(id, type);
     }
 }
 exports.Define = Define;
