@@ -1,53 +1,161 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+const tslib_1 = require("tslib");
 const _ = require("lodash");
-const path = require("path");
 const define_1 = require("./define");
+const decorators_1 = require("./decorators");
+const util_1 = require("./util");
 class Injector {
     constructor() {
         this.FACTORY_POSTFIX = "Factory";
         this._instances = {};
         this._definitions = {};
         this._options = {};
-        this._factories = {};
+        this._factoriesObjects = {};
         this._alias = {};
         this._aliasFactory = {};
+        this._factoriesValues = {};
+        this._factories = [];
+    }
+    get parent() {
+        return this._parent;
+    }
+    set parent(value) {
+        this._parent = value;
     }
     /*
      * public  loads the context by given definitions object
      */
     initialize(options) {
-        this._options = options || {};
-        for (let key in this._options.definitions) {
-            if (this._options.definitions.hasOwnProperty(key)) {
+        return tslib_1.__awaiter(this, void 0, void 0, function* () {
+            this._options = options || {};
+            let keys = Object.keys(this._options.definitions || {});
+            for (let i = 0, len = keys.length; i < len; i++) {
+                let key = keys[i];
                 this._definitions[key] = this._options.definitions[key];
             }
+            yield this._wireObjects();
+        });
+    }
+    _wireObjects() {
+        return tslib_1.__awaiter(this, void 0, void 0, function* () {
+            let keys = Object.keys(this._definitions);
+            for (let i = 0, len = keys.length; i < len; i++) {
+                let objectId = keys[i], definition = this._definitions[objectId];
+                this._initInjectParams(definition);
+                this._initProperties(definition);
+                if (definition.factory) {
+                    this._factories.push(definition.id);
+                }
+                if (definition.aliasFactory) {
+                    this._populateAliasFactory(definition, objectId);
+                }
+            }
+            for (let i = 0, len = keys.length; i < len; i++) {
+                let objectId = keys[i], definition = this._definitions[objectId];
+                (definition.singleton && !definition.lazy) && (this._createObjectInstance(objectId, definition));
+            }
+            keys = Object.keys(this._instances);
+            //loop over instances and inject properties and look up methods only if exist in def
+            for (let i = 0, len = keys.length; i < len; i++) {
+                let objectId = keys[i], instance = this._instances[objectId];
+                (this._definitions[objectId]) && (this._injectPropertiesAndLookUpMethods(instance, this._definitions[objectId], objectId));
+            }
+            yield this._loadFactories(this._factories);
+            for (let i = 0, len = keys.length; i < len; i++) {
+                let objectId = keys[i], instance = this._instances[objectId];
+                if (this._definitions[objectId]) {
+                    this._injectFactoryObject(instance, objectId);
+                    this._injectAlias(this._definitions[objectId], instance);
+                    this._injectAliasFactory(this._definitions[objectId], instance);
+                }
+            }
+            //loop instances and invoke init methods
+            for (let i = 0, len = keys.length; i < len; i++) {
+                let objectId = keys[i], instance = this._instances[objectId];
+                (this._definitions[objectId]) && (this._invokeInitMethod(instance, this._definitions[objectId]));
+            }
+        });
+    }
+    _initInjectParams(def) {
+        let $self = this;
+        if (!def.type) {
+            return;
         }
-        this._wireObjects();
-        return this;
+        let params = Reflect.getMetadata(decorators_1.InjectParamSymbol, def.type);
+        if (!params || !_.isFunction(def.type)) {
+            return;
+        }
+        let paramGroups = _.groupBy(params, "method");
+        _.forEach(paramGroups, (items, method) => {
+            let oldFn = def.type.prototype[method];
+            def.type.prototype[method] = function (...args) {
+                for (let i = 0, length = (items.length || 0); i < length; i++) {
+                    args[items[i].index] = $self.getObject(items[i].param);
+                }
+                return oldFn.apply(this, args);
+            };
+        });
     }
     /*
      * public get object by object Id
      */
-    getObject(objectID, runtimeArgs, ignoreFactory) {
-        return this._get(objectID, runtimeArgs, ignoreFactory);
+    getObject(objectID, runtimeArgs) {
+        return this.get(objectID, runtimeArgs);
     }
     resolve(objectID, runtimeArgs) {
-        return this._get(objectID, runtimeArgs);
+        return this.get(objectID, runtimeArgs);
     }
     get(objectID, runtimeArgs) {
+        objectID = util_1.Util.getClassNameOrId(objectID);
         return this._get(objectID, runtimeArgs);
     }
-    _get(objectID, runtimeArgs, ignoreFactory, referenceChain = []) {
-        //check if we have factory and it's not ignored
-        if (!ignoreFactory && this._definitions[(objectID + this.FACTORY_POSTFIX)]) {
-            return this.getObject(objectID + this.FACTORY_POSTFIX).get();
+    getFactoryValue(objectID) {
+        let def = this._definitions[objectID];
+        if (!def) {
+            return this.parent ? this.parent.getFactoryValue(objectID) : null;
         }
+        if (def.injector) {
+            return def.injector.getFactoryValue(def.id);
+        }
+        return this._factoriesValues[def.id];
+    }
+    getFactory(objectID) {
+        return tslib_1.__awaiter(this, void 0, void 0, function* () {
+            objectID = util_1.Util.getClassNameOrId(objectID);
+            let def = this._definitions[objectID] || this._definitions[objectID + this.FACTORY_POSTFIX];
+            if (!def) {
+                return this.parent ? this.parent.getFactory(objectID) : null;
+            }
+            if (def.injector) {
+                return def.injector.getFactory(def.id);
+            }
+            let value = this._factoriesValues[def.id];
+            if (value) {
+                return value;
+            }
+            let factory = this._get(def.id);
+            this._injectFactoryObject(this._instances[def.id], def.id);
+            value = yield factory.get();
+            this._factoriesValues[def.id] = value;
+            return value;
+        });
+    }
+    _get(objectID, runtimeArgs, referenceChain = []) {
         let instance = this._instances[objectID];
-        if (!instance) {
-            instance = this._createObjectInstance(objectID, this._definitions[objectID], runtimeArgs, referenceChain);
+        if (instance) {
+            return instance;
         }
-        return instance;
+        let definition = this._definitions[objectID];
+        if (definition) {
+            return definition.injector
+                ? definition.injector.getObject(objectID, runtimeArgs)
+                : this._createObjectInstance(objectID, this._definitions[objectID], runtimeArgs, referenceChain);
+        }
+        if (this.parent) {
+            return this.parent.getObject(objectID);
+        }
+        throw new Error(`Injector:can't find object definition for objectID:${objectID} ${referenceChain.join(' -> ')}}`);
     }
     getInstance(objectId) {
         return this._instances[objectId];
@@ -56,6 +164,7 @@ class Injector {
         if (this._definitions[objectId]) {
             console.log(`Injector:definition id already exists overriding:  ${objectId}`);
         }
+        definition.id = objectId;
         this._definitions[objectId] = definition;
         return this;
     }
@@ -76,11 +185,11 @@ class Injector {
         }
         return this;
     }
-    addObject(objectId, instance) {
+    addObject(objectId, instance, silent) {
         return this.addInstance(objectId, instance);
     }
-    addInstance(objectId, instance) {
-        if (this._instances[objectId]) {
+    addInstance(objectId, instance, silent) {
+        if (!silent && this._instances[objectId]) {
             console.log("Injector:object id already exists overriding: " + objectId);
         }
         this._instances[objectId] = instance;
@@ -106,7 +215,13 @@ class Injector {
         return this._definitions;
     }
     getDefinition(id) {
-        return this._definitions[id];
+        let def = this._definitions[id];
+        if (def) {
+            return def.injector ? def.injector.getDefinition(id) : def;
+        }
+        if (this.parent) {
+            return this.parent.getDefinition(id);
+        }
     }
     getAlias(aliasName) {
         return this._alias[aliasName] || [];
@@ -117,42 +232,46 @@ class Injector {
     delegate(objectId) {
         let self = this;
         return function () {
-            let object = self.getObject(objectId);
+            let object = self._get(objectId);
             object.run.apply(object, arguments);
         };
     }
-    define(id, type) {
-        return new define_1.Define(this, id, type);
+    registerMulti(fns) {
+        for (let i = 0, len = fns.length; i < len; i++) {
+            this.register(fns[i]);
+        }
+        return this;
     }
-    reset() {
-        this._instances = {};
-        this._definitions = {};
-        this._options = {};
-        this._alias = {};
-        this._aliasFactory = {};
-        this._factories = {};
+    register(id, type) {
+        if (_.isFunction(id)) {
+            type = id;
+            id = util_1.Util.getClassName(type);
+        }
+        let define = Reflect.getMetadata(decorators_1.InjectDefineSymbol, type) || new define_1.Define(id, type);
+        this.addDefinition(id, define.definition);
+        return define;
     }
     _createObjectInstance(objectID, objectDefinition, runtimeArgs, referenceChain = []) {
         let argumentInstances = [], args, newObjectInstance;
         //check Circular reference
-        if (referenceChain.indexOf(objectID) > -1) {
+        if (referenceChain.length && referenceChain.indexOf(objectID) > -1) {
             referenceChain.push(objectID);
             throw new Error(`Circular reference ${referenceChain.join(' -> ')}`);
         }
         referenceChain.push(objectID);
         //checks if we have a valid object definition
         if (!objectDefinition) {
-            throw new Error("Injector:can't find object definition for objectID:" + objectID);
+            throw new Error(`Injector:can't find object definition for objectID:${objectID} ${referenceChain.join(' -> ')}}`);
         }
         newObjectInstance = this._instances[objectID];
         //	If the instance does not already exist make it
         if (newObjectInstance) {
             return newObjectInstance;
         }
-        //convert path to type
-        if (objectDefinition.path) {
-            objectDefinition.type = require(path.join(this._options.root, objectDefinition.path + '.js'));
-        }
+        // //convert path to type
+        // if (objectDefinition.path) {
+        //     objectDefinition.type = require(path.join(this._options.root, objectDefinition.path + '.js'))
+        // }
         args = (objectDefinition.args && objectDefinition.args.length > 0) ? objectDefinition.args : [];
         //add runtime args to the end of args obj
         for (let i = 0, length = (runtimeArgs ? runtimeArgs.length : 0); i < length; i++) {
@@ -161,7 +280,7 @@ class Injector {
         //loop over args and get the arg value or create arg object instance
         for (let i = 0, length = (args ? args.length : 0); i < length; i++) {
             let arg = args[i];
-            argumentInstances.push(arg.hasOwnProperty("value") ? arg.value : this._get(arg.ref, [], false, referenceChain));
+            argumentInstances.push(arg.hasOwnProperty("value") ? arg.value : this._get(arg.ref, [], referenceChain));
         }
         try {
             //crate object instance
@@ -185,53 +304,12 @@ class Injector {
     /*
      * private creates new objects instances and inject properties
      */
-    _wireObjects() {
-        //add aliasFactory
-        for (let objectId in this._definitions) {
-            if (this._definitions.hasOwnProperty(objectId)) {
-                let definition = this._definitions[objectId];
-                if (definition.aliasFactory) {
-                    this._populateAliasFactory(definition, objectId);
-                }
-            }
-        }
-        for (let objectId in this._definitions) {
-            if (this._definitions.hasOwnProperty(objectId)) {
-                let definition = this._definitions[objectId];
-                (definition.singleton && !definition.lazy) && (this._createObjectInstance(objectId, definition));
-            }
-        }
-        //loop over instances and inject properties and look up methods only if exist in def
-        for (let objectId in this._instances) {
-            if (this._instances.hasOwnProperty(objectId)) {
-                let instance = this._instances[objectId];
-                (this._definitions[objectId]) && (this._injectPropertiesAndLookUpMethods(instance, this._definitions[objectId], objectId));
-            }
-        }
-        for (let objectId in this._instances) {
-            if (this._instances.hasOwnProperty(objectId)) {
-                let instance = this._instances[objectId];
-                if (this._definitions[objectId]) {
-                    this._injectFactoryObject(instance, objectId);
-                    this._injectAlias(this._definitions[objectId], instance);
-                    this._injectAliasFactory(this._definitions[objectId], instance);
-                }
-            }
-        }
-        //loop instances and invoke init methods
-        for (let objectId in this._instances) {
-            if (this._instances.hasOwnProperty(objectId)) {
-                let instance = this._instances[objectId];
-                (this._definitions[objectId]) && (this._invokeInitMethod(instance, this._definitions[objectId]));
-            }
-        }
-    }
     _populateAliasFactory(definition, objectId) {
         for (let i = 0, length = definition.aliasFactory ? definition.aliasFactory.length : 0; i < length; i++) {
             let aliasName = definition.aliasFactory[i];
-            let delegateFn = this._createDelegate(this.getObject, this, [objectId]);
+            let delegateFn = util_1.Util.createDelegate(this._get, this, [objectId]);
             delegateFn.type = definition.type;
-            this._mapPush(this._aliasFactory, aliasName, delegateFn);
+            util_1.Util.mapPush(this._aliasFactory, aliasName, delegateFn);
         }
     }
     /*
@@ -242,20 +320,34 @@ class Injector {
             object[definition.initMethod]();
         }
     }
-    _getProperties(objectDefinition) {
-        let properties = objectDefinition.props || objectDefinition.properties || [];
-        if (!objectDefinition.$propertiesGenerated) {
-            for (let i = 0, length = (objectDefinition.inject ? objectDefinition.inject.length : 0); i < length; i++) {
-                let injectable = objectDefinition.inject[i];
-                properties.push(_.isString(injectable) ? ({
+    _initProperties(definition) {
+        // if (definition.factory) {
+        //     this._factories.push(definition.id);
+        // }
+        let properties = definition.props || definition.properties || [];
+        for (let i = 0, length = (definition.inject ? definition.inject.length : 0); i < length; i++) {
+            let injectable = definition.inject[i];
+            let dto = injectable;
+            if (_.isString(injectable)) {
+                dto = {
                     name: injectable,
                     ref: injectable
-                }) : injectable);
+                };
             }
-            objectDefinition.properties = properties;
-            objectDefinition.$propertiesGenerated = true;
+            if (dto.ref) {
+                let factoryRef = dto.ref + this.FACTORY_POSTFIX, factoryDef = this.getDefinition(factoryRef), refDef = this.getDefinition(dto.ref);
+                if (refDef && refDef.factory) {
+                    dto.factory = refDef.id;
+                    delete dto.ref;
+                }
+                else if (factoryDef && factoryDef.factory && definition.id != factoryRef) {
+                    dto.factory = factoryRef;
+                    delete dto.ref;
+                }
+            }
+            properties.push(dto);
         }
-        return properties;
+        definition.properties = properties;
     }
     /*
      * private  fire single object instance
@@ -274,44 +366,37 @@ class Injector {
      * private inject values and look up methods to object properties
      */
     _injectPropertiesAndLookUpMethods(object, objectDefinition, objectId) {
-        let injectObject, obj, factoryRef, properties = this._getProperties(objectDefinition);
+        let injectObject, obj, properties = objectDefinition.properties;
         //loop over the properties definition
         for (let i = 0, length = (properties ? properties.length : 0); i < length; i++) {
             let prop = properties[i];
             //get property obj
             injectObject = null;
-            factoryRef = prop.ref + this.FACTORY_POSTFIX;
             if (prop.array) {
-                injectObject = _.map(prop.array, (propObj) => propObj.value || this.getObject(propObj.ref));
+                injectObject = _.map(prop.array, (propObj) => propObj.value || this._get(propObj.ref));
             }
             else if (prop.dictionary) {
                 injectObject = {};
-                _.forEach(prop.dictionary, (propObj) => injectObject[propObj.key] = propObj.value || this.getObject(propObj.ref));
+                _.forEach(prop.dictionary, (propObj) => injectObject[propObj.key] = propObj.value || this._get(propObj.ref));
             }
             else if (prop.value) {
                 injectObject = prop.value;
             }
-            else if (prop.ref && !this._definitions[factoryRef]) {
-                injectObject = this.getObject(prop.ref);
+            else if (prop.ref) { //check if we have ref and we don't have factory with the same name
+                injectObject = this._get(prop.ref);
             }
             else if (prop.objectProperty) {
-                obj = this.getObject(prop.objectProperty.object);
+                obj = this._get(prop.objectProperty.object);
                 injectObject = obj[prop.objectProperty.property];
             }
-            else if (prop.factory || this._definitions[factoryRef]) {
-                if (factoryRef == objectId) {
-                    injectObject = this._get(prop.ref, [], true);
+            else if (prop.factory) {
+                if (!this._factoriesObjects[objectId]) {
+                    this._factoriesObjects[objectId] = {};
                 }
-                else {
-                    let factoryName = prop.factory || factoryRef;
-                    if (!this._factories[objectId]) {
-                        this._factories[objectId] = {};
-                    }
-                    this._factories[objectId][prop.name] = factoryName;
-                }
+                this._factoriesObjects[objectId][prop.name] = prop.factory;
             }
             else if (prop.factoryMethod) {
-                injectObject = this._createDelegate(this.getObject, this, [prop.factoryMethod]);
+                injectObject = util_1.Util.createDelegate(this._get, this, [prop.factoryMethod]);
             }
             if (injectObject) {
                 object[prop.name] = injectObject;
@@ -322,24 +407,53 @@ class Injector {
         }
         //add alias
         if (objectDefinition.alias && objectDefinition.singleton) {
-            for (let key in objectDefinition.alias) {
-                if (objectDefinition.alias.hasOwnProperty(key)) {
-                    this._mapPush(this._alias, objectDefinition.alias[key], object);
-                }
+            let keys = Object.keys(objectDefinition.alias);
+            for (let i = 0, len = keys.length; i < len; i++) {
+                let key = keys[i];
+                util_1.Util.mapPush(this._alias, objectDefinition.alias[key], object);
             }
         }
+    }
+    _loadFactories(factories) {
+        return tslib_1.__awaiter(this, void 0, void 0, function* () {
+            for (let factory of factories) {
+                yield this.loadFactory(factory);
+            }
+        });
+    }
+    loadFactory(objectId) {
+        return tslib_1.__awaiter(this, void 0, void 0, function* () {
+            let factoryData = this._factoriesObjects[objectId];
+            let keys = Object.keys(factoryData || {});
+            for (let propName of keys) {
+                let factory = factoryData[propName];
+                yield this.loadFactory(factory);
+            }
+            this._factoriesValues[objectId] = yield this.getFactory(objectId);
+        });
     }
     _injectFactoryObject(object, objectId) {
-        let factoryData = this._factories[objectId];
-        if (factoryData) {
-            for (let propName in factoryData) {
-                if (factoryData.hasOwnProperty(propName)) {
-                    let factory = factoryData[propName];
-                    object[propName] = this.getObject(factory).get();
-                }
-            }
+        let factoryData = this._factoriesObjects[objectId];
+        //let objectDef = this._definitions[objectId];
+        if (!factoryData) {
+            return;
         }
+        let keys = Object.keys(factoryData);
+        for (let i = 0, len = keys.length; i < len; i++) {
+            let propName = keys[i], factory = factoryData[propName];
+            //let factoryDef = this._definitions[factory];
+            //this._injectFactoryObject(this._instances[factoryDef.id], factoryDef.id);
+            //let factoryValue = this._factoriesValues[factory];
+            // if(!factoryValue){
+            //    factoryValue = this._factoriesValues[factory] =  this._get<IFactory<T>>(factory).get();
+            //}
+            object[propName] = this.getFactoryValue(factory);
+        }
+        //objectDef.$isFactoryWired = true;
     }
+    // private _injectFactoryObjectSync() {
+    //
+    // }
     _injectAlias(definition, instance) {
         for (let i = 0, length = (definition.properties ? definition.properties.length : 0); i < length; i++) {
             let prop = definition.properties[i];
@@ -352,19 +466,13 @@ class Injector {
             (prop.aliasFactory) && (instance[prop.name] = prop.indexBy ? _.keyBy(this.getAliasFactory(prop.aliasFactory), (item) => item.type[prop.indexBy]) : this.getAliasFactory(prop.aliasFactory));
         }
     }
-    _createDelegate(fn, obj, args) {
-        return function () {
-            let callArgs = (args || []).concat(arguments);
-            return fn.apply(obj, callArgs);
-        };
-    }
-    _mapPush(map, key, obj) {
-        (!map[key]) && (map[key] = []);
-        map[key].push(obj);
-    }
-    _getFunctionArgs(func) {
-        let STRIP_COMMENTS = /((\/\/.*$)|(\/\*[\s\S]*?\*\/))/mg, ARGUMENT_NAMES = /([^\s,]+)/g, fnStr = func.toString().replace(STRIP_COMMENTS, '').replace(/(\r\n|\n|\r|\t| )/gm, ""), args = fnStr.slice(fnStr.indexOf('constructor(') + 12, fnStr.indexOf(')')).match(ARGUMENT_NAMES) || [];
-        return _.compact(args);
+    reset() {
+        this._instances = {};
+        this._definitions = {};
+        this._options = {};
+        this._alias = {};
+        this._aliasFactory = {};
+        this._factoriesObjects = {};
     }
 }
 exports.Injector = Injector;
