@@ -1,43 +1,46 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+const tslib_1 = require("tslib");
 const chai = require("chai");
-const inject = require("../lib/inject");
+const ioc = require("../lib/inject");
 let should = chai.should();
 describe('Alias Factory', function () {
     describe('should inject alias factory', function () {
         let injector, CalcManager;
         beforeEach(function () {
-            injector = inject.createContainer();
-            let Rectangle = class {
-                constructor() {
-                }
-            };
-            CalcManager = class {
-                constructor(num) {
-                    this.num = num || 25;
-                }
-                calc() {
-                    return this.num;
-                }
-            };
-            injector.addDefinitions({
-                rectangle: {
-                    type: Rectangle,
-                    singleton: true,
-                    props: [
-                        {
-                            name: 'calcable',
-                            aliasFactory: 'calcable'
-                        }
-                    ]
-                },
-                calcManager: {
-                    type: CalcManager,
-                    aliasFactory: ['calcable'],
-                    singleton: false
-                }
+            return tslib_1.__awaiter(this, void 0, void 0, function* () {
+                injector = ioc.createContainer();
+                let Rectangle = class {
+                    constructor() {
+                    }
+                };
+                CalcManager = class {
+                    constructor(num) {
+                        this.num = num || 25;
+                    }
+                    calc() {
+                        return this.num;
+                    }
+                };
+                injector.addDefinitions({
+                    rectangle: {
+                        type: Rectangle,
+                        singleton: true,
+                        props: [
+                            {
+                                name: 'calcable',
+                                aliasFactory: 'calcable'
+                            }
+                        ]
+                    },
+                    calcManager: {
+                        type: CalcManager,
+                        aliasFactory: ['calcable'],
+                        singleton: false
+                    }
+                });
+                yield injector.initialize();
             });
-            injector.initialize();
         });
         it('should inject property ', function () {
             let rectangle = injector.getObject('rectangle');
@@ -45,7 +48,7 @@ describe('Alias Factory', function () {
             rectangle.calcable.should.be.an.instanceOf(Array);
             rectangle.calcable.length.should.be.equal(1);
             let calcable = rectangle.calcable[0]();
-            calcable.calc.should.be.a.Function;
+            calcable.calc.should.be.a('function');
             calcable.calc().should.be.eq(25);
         });
         it('should inject property with run time params', function () {
@@ -54,31 +57,33 @@ describe('Alias Factory', function () {
             rectangle.calcable.should.be.an.instanceOf(Array);
             rectangle.calcable.length.should.be.equal(1);
             let calcable = rectangle.calcable[0](30);
-            calcable.calc.should.be.a.Function;
+            calcable.calc.should.be.a('function');
             calcable.calc().should.be.eq(30);
         });
     });
     describe('should inject alias factory link', function () {
         let injector, CalcManager;
         beforeEach(function () {
-            injector = inject.createContainer();
-            let Rectangle = class {
-                constructor() {
-                }
-            };
-            CalcManager = class {
-                constructor(num) {
-                    this.num = num || 25;
-                }
-                calc() {
-                    return this.num;
-                }
-            };
-            injector.define('rectangle', Rectangle)
-                .singleton()
-                .injectAliasFactory('calcable', 'calcable');
-            injector.define('calcManager', CalcManager).aliasFactory(['calcable']);
-            injector.initialize();
+            return tslib_1.__awaiter(this, void 0, void 0, function* () {
+                injector = ioc.createContainer();
+                let Rectangle = class {
+                    constructor() {
+                    }
+                };
+                CalcManager = class {
+                    constructor(num) {
+                        this.num = num || 25;
+                    }
+                    calc() {
+                        return this.num;
+                    }
+                };
+                injector.register('rectangle', Rectangle)
+                    .singleton()
+                    .injectAliasFactory('calcable', 'calcable');
+                injector.register('calcManager', CalcManager).aliasFactory(['calcable']);
+                yield injector.initialize();
+            });
         });
         it('should inject property ', function () {
             var rectangle = injector.getObject('rectangle');
@@ -86,7 +91,7 @@ describe('Alias Factory', function () {
             rectangle.calcable.should.be.an.instanceOf(Array);
             rectangle.calcable.length.should.be.equal(1);
             let calcable = rectangle.calcable[0]();
-            calcable.calc.should.be.a.Function;
+            calcable.calc.should.be.a('function');
             calcable.calc().should.be.eq(25);
         });
         it('should inject property with run time params', function () {
@@ -95,39 +100,41 @@ describe('Alias Factory', function () {
             rectangle.calcable.should.be.an.instanceOf(Array);
             rectangle.calcable.length.should.be.equal(1);
             let calcable = rectangle.calcable[0](30);
-            calcable.calc.should.be.a.Function;
+            calcable.calc.should.be.a('function');
             calcable.calc().should.be.eq(30);
         });
     });
     describe('should inject alias factory link indexBy', function () {
         let injector, CalcManager;
         beforeEach(function () {
-            injector = inject.createContainer();
-            let Rectangle = class {
-                constructor() {
-                }
-            };
-            CalcManager = class {
-                static get NAME() {
-                    return "test";
-                }
-                constructor(num) {
-                    this.num = num || 25;
-                }
-                calc() {
-                    return this.num;
-                }
-            };
-            injector.define('rectangle', Rectangle).singleton().injectAliasFactory('calcable', 'calcable', "NAME");
-            injector.define('calcManager', CalcManager).aliasFactory(['calcable']);
-            injector.initialize();
+            return tslib_1.__awaiter(this, void 0, void 0, function* () {
+                injector = ioc.createContainer();
+                let Rectangle = class {
+                    constructor() {
+                    }
+                };
+                CalcManager = class {
+                    static get NAME() {
+                        return "test";
+                    }
+                    constructor(num) {
+                        this.num = num || 25;
+                    }
+                    calc() {
+                        return this.num;
+                    }
+                };
+                injector.register('rectangle', Rectangle).singleton().injectAliasFactory('calcable', 'calcable', "NAME");
+                injector.register('calcManager', CalcManager).aliasFactory(['calcable']);
+                yield injector.initialize();
+            });
         });
         it('should inject property ', function () {
             let rectangle = injector.getObject('rectangle');
             should.exist(rectangle.calcable);
             rectangle.calcable.should.be.an.instanceOf(Object);
             let calcable = rectangle.calcable.test();
-            calcable.calc.should.be.a.Function;
+            calcable.calc.should.be.a('function');
             calcable.calc().should.be.eq(25);
         });
     });
