@@ -1,11 +1,13 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-let should = require('chai').should(), inject = require('../lib/inject');
+const ioc = require("../lib/inject");
+const chai = require("chai");
+let should = chai.should();
 describe('Property Value', function () {
     describe('inject value to object', function () {
         let injector;
         beforeEach(function () {
-            injector = inject.createContainer();
+            injector = ioc.createContainer();
             class Rectangle {
                 constructor() {
                     this.number = Math.random();
@@ -35,7 +37,7 @@ describe('Property Value', function () {
     describe('inject value to object linq', function () {
         let injector;
         it('should have the injected value', function () {
-            injector = inject.createContainer();
+            injector = ioc.createContainer();
             class Rectangle {
                 constructor() {
                     this.number = Math.random();
@@ -44,7 +46,7 @@ describe('Property Value', function () {
                     return this.size;
                 }
             }
-            injector.define('rectangle', Rectangle).singleton().injectValue('size', 25);
+            injector.register('rectangle', Rectangle).singleton().injectValue('size', 25);
             injector.initialize();
             let rectangle = injector.getObject('rectangle');
             should.exist(rectangle.size);
